Move package multi to package channel
