+++ channel/multi/adjudicator.go
