+++ channel/multi/asset.go
