+++ watcher/local/watcher.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/pkg/errors"
 
-	"perun.network/go-perun/backend/multi"
 	"perun.network/go-perun/channel"
+	"perun.network/go-perun/channel/multi"
 	"perun.network/go-perun/log"
 	"perun.network/go-perun/watcher"
 )
